ETL_pipeline/requirements.txt
paho-mqtt==1.6.1
pandas==2.2.2
fastparquet==2024.5.0
pyarrow==16.1.0 ; python_version < "3.13"
streamlit>=1.33
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.1
scikit-learn
